pyproject.toml
[project]
name = "vaultmind-ai"
version = "0.1.0"
description = "FastAPI service for VaultMind AI/RAG processing"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "pydantic-settings>=2.0,<3.0",
    "python-dotenv>=1.0,<2.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]